service interfaces and implementation separated
