+++ src/main/java/com/github/minigithub/model/State.java
@@ -2,6 +2,6 @@ package com.github.minigithub.model;
 
 public enum State {
    OPEN,
-   CLOSE,
+   CLOSED,
    MERGED;
 }
